requirements.txt
fastapi
uvicorn
python-dotenv
httpx
requests
ic-py
pynacl
chromadb
openai
pydantic
